Move expert_evaluation
