+++ athena/playground/src/components/view_mode/evaluation_mode/index.tsx
@@ -6,7 +6,7 @@ import { useState } from "react";
 import DefineExperiment from "./define_experiment";
 import ConfigureModules from "./configure_modules";
 import ConductExperiment from "./conduct_experiment";
-import EvaluationManagement from "@/components/view_mode/evaluation_mode/expert_evaluation/evaluation_management";
+import EvaluationManagement from "@/components/view_mode/expert_evaluation/evaluation_management";
 
 export default function EvaluationMode() {
   const [experiment, setExperiment] = useState<Experiment | undefined>(
